refactor: introduce honor utility class
